+++ PC-kom-src/src/Frame_imp/utils_Pazur/Messages.hpp
@@ -10,7 +10,7 @@ class Messages
 {
 public:
     Messages(QByteArray dat, int siz, char addcrc, uint id);
-    Messages(QList<Message> msgs, uchar id);
+    Messages(QList<Message> msgs, uint id);
     Messages();
 
     QString toQString() const;
